cli: move `workspace` code into `workspace.rs`
Summary: Following in the footsteps of Antoine as I prepare to add multi-
revision support to `workspace add`.
